+++ packages/expo-location/android/src/main/java/expo/modules/location/services/LocationTaskService.java
@@ -20,8 +20,8 @@ import android.util.Log;
 public class LocationTaskService extends Service {
   private static final String TAG = "LocationTaskService";
   private static int sServiceId = 481756;
-
   private String mChannelId;
+  private String killService = "false";
   private Context mParentContext;
   private int mServiceId = sServiceId++;
   private final IBinder mBinder = new ServiceBinder();
@@ -64,8 +64,13 @@ public class LocationTaskService extends Service {
 
   @Override
   public void onTaskRemoved(Intent rootIntent) {
-    super.onTaskRemoved(rootIntent);
-    stop();
+    Bundle extras = rootIntent.getExtras();
+    extras.getString("killService");
+    
+    if(killService == "true"){
+      super.onTaskRemoved(rootIntent);
+      stop();
+    }
   }
 
   public void startForeground(Bundle serviceOptions) {