+++ packages/expo-location/src/Location.types.ts
@@ -203,6 +203,10 @@ export type LocationTaskServiceOptions = {
    * Color of the foreground service notification. Accepts `#RRGGBB` and `#AARRGGBB` hex formats.
    */
   notificationColor?: string;
+  /**
+   * Boolean value whether to destroy the foreground service if the app is killed.
+   */
+  killServiceOnDestroy?: boolean;
 };
 
 // @needsAudit