+++ packages/expo-location/CHANGELOG.md
@@ -12,6 +12,11 @@
 
 ### 💡 Others
 
+## Unpublished
+
+### 🛠 Breaking changes
+
+- Add an option to whether kill or keep the foreground service when app is killed on Android. ([#15633](https://github.com/expo/expo/pull/15633) by [@islamouzou](https://github.com/islamouzou))
 - Updated `@expo/config-plugins` from `4.0.2` to `4.0.14` ([#15621](https://github.com/expo/expo/pull/15621) by [@EvanBacon](https://github.com/EvanBacon))
 
 ## 14.0.1 — 2021-12-15